requirements.txt
pdfminer.six
python-docx
rapidfuzz
nltk
